requirements.txt
# matplotlib
# pyqt5
# qimage2ndarray
# scikit-image>=0.15.0
# scipy
pyserial
numpy